+++ c8y_api/app/__init__.py
@@ -3,10 +3,13 @@
 # and/or its subsidiaries and/or its affiliates and/or their licensors.
 # Use, reproduction, transfer, publication or disclosure is prohibited except
 # as specifically provided for in your License Agreement with Software AG.
-
+import base64
 import dataclasses
+import json
 import logging
 import os
+
+from functools import lru_cache
 
 from c8y_api._base_api import c8y_keys
 from c8y_api._main_api import CumulocityApi
@@ -129,16 +132,65 @@ class CumulocityApp(CumulocityApi):
         return cls._bootstrap_instance
 
     @classmethod
-    def get_tenant_instance(cls, tenant_id) -> CumulocityApi:
+    def get_tenant_instance(cls, tenant_id: str = None, headers: dict = None) -> CumulocityApi:
         """Provide access to a tenant-specific instance in a multi-tenant
         application setup.
 
         Args:
             tenant_id (str):  ID of the tenant to get access to
+            headers (dict):  Inbound request headers, the tenant ID
+                is resolved from the Authorization header
 
         Returns:
             A CumulocityApi instance authorized for a tenant user
         """
+        # (1) if the tenant ID is specified we just
+        if tenant_id:
+            return cls._get_tenant_instance(tenant_id)
+
+        # (2) otherwise, look for the Authorization header
+        if not headers:
+            raise RuntimeError("At least one of 'tenant_id' or 'headers' must be specified.")
+
+        auth_header = headers[next(filter(lambda k: 'Authorization'.upper() == k.upper(), headers.keys()))]
+        if not auth_header:
+            raise ValueError("Missing Authentication header. Unable to resolve tenant ID.")
+
+        return cls._get_tenant_instance(cls._resolve_tenant_id_from_auth_header(auth_header))
+
+    @classmethod
+    def _get_tenant_instance(cls, tenant_id: str) -> CumulocityApi:
         if tenant_id not in cls._tenant_instances:
             cls._tenant_instances[tenant_id] = CumulocityApp(tenant_id)
         return cls._tenant_instances[tenant_id]
+
+    @classmethod
+    @lru_cache(maxsize=128, typed=False)
+    def _resolve_tenant_id_from_auth_header(cls, auth_header: str) -> str:
+        auth_type, auth_value = auth_header.split(' ')
+
+        if auth_type.upper() == 'BASIC':
+            return cls._resolve_tenant_id_basic(auth_value)
+        if auth_type.upper() == 'BEARER':
+            return cls._resolve_tenant_id_token(auth_value)
+
+        raise ValueError(f"Unexpected authorization header type: {auth_type}")
+
+    @staticmethod
+    def _resolve_tenant_id_basic(auth_string: str) -> str:
+        decoded = base64.b64decode(bytes(auth_string, 'utf-8'))
+        username = decoded.split(b':', 1)[0]
+        if b'/' not in username:
+            raise ValueError(f"nable to resolve tenant ID. Username '{username}' does not appear to include it.")
+        return username.split(b'/', 1)[0].decode('utf-8')
+
+    @classmethod
+    def _resolve_tenant_id_token(cls, auth_token: str) -> str:
+        # we assume that the token is an JWT token
+        jwt_parts = auth_token.split('.')
+        if len(jwt_parts) != 3:
+            raise ValueError("Unexpected token format (not an JWT?). Unable to resolve tenant ID.")
+        jwt_body = json.loads(base64.b64decode(jwt_parts[1].encode('utf-8')))
+        if 'ten' not in jwt_body:
+            raise ValueError("Unexpected token format (missing 'ten' claim). Unable to resolve tenant ID.")
+        return jwt_body['ten']