+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # Changelog
 
+
 ## Work in progress
 
 ### Changed
@@ -9,12 +10,16 @@
 * Fixed [Issue #7](https://github.com/SoftwareAG/cumulocity-python-api/issues/7):
   Improved caching and user experience when creating CumulocityApp instances. Added unit tests.
 
+* Added possibility to resolve the tenant ID from authorization headers (both `Basic` and `Bearer`).
+
+
 ## Version 1.0.1
 
 ### Changed
 
 * The cumulocity-pyton-api library is now available on [PyPI](https://pypi.org) under the name `c8y_api` (see https://pypi.org/project/c8y-api/) 
 * Updated README to reflect installation from PyPI 
+
 
 ## Version 1.0